RavenDB-16880 Major refactor - fix bug where connection ends after non fatal error
The reason for this refactor is that we needed to be able to get an instance of the message even if continuing reading it will cause a PgErrorException.
We needed the instance to later do message.HandleError which will result in different behaviour for ExtendedProtocolMessage types (set transaction state to failed) and regular Message types.